requirements.txt
# Core dependencies for LangChain + LangSmith observability
langchain>=0.3.0,<2.0.0
langchain-openai>=0.2.0,<2.0.0
langsmith>=0.1.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
